README: Fix the broken link
This document has a link to a document that shall be located in the subfolder `./Timestamping/docs`. However, the referred file has changed its name. Hence, the link is broken. It also uses an absolute link to a specific version, which also a bit problematic. It'd better for it to have a relative link as advised in this commit.